pyproject.toml
[project]
name = "pii-safe-poc"
version = "0.1.0"
description = "Privacy middleware PoC — PII detection, policy-driven sanitization, and MCP tool interface"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "spacy>=3.7",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
pii-safe = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
